[FIXME] sdfat: update to version 2.4.5 from G980FXXU5CTKG.
After this update, there are kernel panics caused by sdfat, when writing
heavily to a filesystem mounted with it. I've been able to get these
kernel panics when recording videos or downloading torrents.
TODO: find the part of this update causing this issue and revert it, or
find a proper fix.